Android: Permissions for showLibraryButton
The showLibraryButton code path needs to be updated to properly handle Android 6 runtime permissions. This basically uses the same flow as for takePicture since the same permissions are required.

It's a little aggressive right now, but it's the best we can do in the short timeframe and given the implementation in the camera plugin.

This also includes the addition of a new dialog which is shown if permissions are denied, and allows the user to link to the settings app.