iOS: Don't save library image to library
With the changes to allow choosing a library image from within the camera UI, it exposed a small bug whereby the library image could be re-written out to the library. This creates duplicate images; this fix adjusts the behavior to only do that if the image source is the camera.