iOS: Update Library Button placement
Updates the placement of the Library button on iPad to better match the location used on iPhone.

This works better now that multitasking is supported -- in Split View, the camera controls change to be more like the iPhone controls, so the consistent location of the Library buttons allows everything to lay out more cleanly.