Android: Fix permission issue with WRITE_EXTERNAL_STORAGE
Consolidate the code to check/request permissions for using the camera to a single function that checks for both read and write storage access. Also adds some addtional debug logging.

Previously, we had similar, but slightly different, logic for checking permissions in the take a picture vs show chooser case. In the latter, we only requested READ permissions if we already had CAMERA permission, when in fact we needed WRITE permission as well if we are saving the photo to the library. By updating the permission check logic to be the same as for the take a picture case, we ensure we have the right permissions.